Use corpus to generate tags
Avoid re-parsing the raw text when we tag the document.

We're not tagging any other doucuments once the model has been trained,
so we can just use the corpus we've already created, which can be saved
and loaded between experiments.

Store the metadata about a document so we can use it to generate the
tags, even when we save and load the corpus.