Filter out much more stopwords
Filter out more unigrams and bigrams by combining multiple stopword lists.

We have quite a high bar for things we consider terms, because the terms
that end up in a topic need to relate to the actual content.

Previously we were using gensim's stopwords in one case and nltk's in
another. I've dropped the nltk one entirely, but added in Matt's
suggested list from
https://github.com/ivyleavedtoadflax/easyNLP/blob/master/sentiment_analysis/stopwords.txt
as `stopwords/easyNLP.txt`.

We found that there are a lot of words that just describe websites and
documents and processes, and these clutter up the output.
`stopwords/custom.txt` and `stopwords/extensions.txt` contain these.

We also have a lot of names and titles (for example in speeches and
news) which don't usually tell you much about the content itself.  I've
created `stopwords/names.txt` and `stopwords/title.txt` from the list of
people on GOV.UK:
https://www.gov.uk/api/search.json?filter_format=person

Finally, `stopwords/time.txt` excludes weekdays and month names.