Simplify CLI options
Previously, you had to specify some combination of `--output-tags`,
`--output-topics`, and `--vis-filename` to run `train_lda.py`, otherwise it
wouldn't generate any output.

This was really annoying, and none of these things are expensive to
generate, so now we just always generate them, and the filenames are all
optional.

`--use-lemmatisation` is now on by default, so you need to pass
`no-lemmatiation` instead to disable it.

This means the thing can run with less options, eg:

    python train_lda.py --experiment foo input/foo.csv

or:

    python train_lda.py --numtopics 15 refine foo

By default, everything is now saved to a subdirectory of `experiment/`.
When running an `import`, you can specify the name of this with
`--experiment`.